+++ src/Table/index.tsx
@@ -18,30 +18,30 @@
  */
 
 import {
-  ColumnDef,
-  flexRender,
-  getCoreRowModel,
-  getPaginationRowModel,
-  getSortedRowModel,
-  useReactTable,
+	ColumnDef,
+	flexRender,
+	getCoreRowModel,
+	getPaginationRowModel,
+	getSortedRowModel,
+	useReactTable,
 } from '@tanstack/react-table';
-import {
-  Loader,
-  Resizer,
-  SortButton,
-  TableStyled,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableHeader,
-  TableRow,
-  TableCellWrapper,
-  TableHeaderWrapper,
-  TableContainerInner,
-} from './styled';
 import { TablePaginationV8 } from './TablePagination';
 import { TableTabs, TableTabsHandler, TableTabsInput } from './TableTabs';
+import {
+	Loader,
+	Resizer,
+	SortButton,
+	TableBody,
+	TableCell,
+	TableCellWrapper,
+	TableContainer,
+	TableContainerInner,
+	TableHead,
+	TableHeader,
+	TableHeaderWrapper,
+	TableRow,
+	TableStyled,
+} from './styled';
 import { ReactTableCustomProps } from './types';
 
 // IMPORTANT
@@ -49,179 +49,179 @@ import { ReactTableCustomProps } from './types';
 // see the readme & stories for help.
 
 declare module '@tanstack/table-core' {
-  interface ColumnMeta<TData extends unknown, TValue> {
-    columnTabs?: {
-      activeTab: string;
-      handleTabs: TableTabsHandler;
-      tabs: TableTabsInput;
-    };
-    customCell?: boolean;
-    customHeader?: boolean;
-    multiSortIds?: string[];
-  }
+	interface ColumnMeta<TData extends unknown, TValue> {
+		columnTabs?: {
+			activeTab: string;
+			handleTabs: TableTabsHandler;
+			tabs: TableTabsInput;
+		};
+		customCell?: boolean;
+		customHeader?: boolean;
+		multiSortIds?: string[];
+	}
 }
 
-interface ReactTableProps<TData> extends ReactTableCustomProps {
-  columns?: ColumnDef<TData>[];
-  data?: TData[];
+export interface ReactTableProps<TData> extends ReactTableCustomProps {
+	columns?: ColumnDef<TData>[];
+	data?: TData[];
 }
 
 export const DEFAULT_TABLE_PAGE_SIZE = 20;
 
 // if not using pagination, put all rows on one page.
 const singlePagePaginationState = {
-  pagination: { pageIndex: 0, pageSize: Number.MAX_SAFE_INTEGER },
+	pagination: { pageIndex: 0, pageSize: Number.MAX_SAFE_INTEGER },
 };
 
 export const Table = <TData extends object>({
-  className = '',
-  columns = [],
-  data = [],
-  enableColumnResizing = false,
-  enableSorting = false,
-  initialState = {},
-  LoaderComponent = Loader,
-  loading = false,
-  manualPagination = false,
-  manualSorting = false,
-  onPaginationChange,
-  onSortingChange,
-  pageCount,
-  paginationState = null,
-  showPageSizeOptions = false,
-  sortingState = null,
-  withFilters = false,
-  withHeaders = false,
-  withPagination = false,
-  withRowBorder = false,
-  withRowHighlight = false,
-  withSideBorders = false,
-  withStripes = false,
-  withTabs = false,
+	className = '',
+	columns = [],
+	data = [],
+	enableColumnResizing = false,
+	enableSorting = false,
+	initialState = {},
+	LoaderComponent = Loader,
+	loading = false,
+	manualPagination = false,
+	manualSorting = false,
+	onPaginationChange,
+	onSortingChange,
+	pageCount,
+	paginationState = null,
+	showPageSizeOptions = false,
+	sortingState = null,
+	withFilters = false,
+	withHeaders = false,
+	withPagination = false,
+	withRowBorder = false,
+	withRowHighlight = false,
+	withSideBorders = false,
+	withStripes = false,
+	withTabs = false,
 }: ReactTableProps<TData>) => {
-  const reactTable = useReactTable({
-    columnResizeMode: 'onChange',
-    columns,
-    data,
-    enableColumnResizing,
-    enableSorting,
-    getCoreRowModel: getCoreRowModel(),
-    initialState: { pagination: { pageSize: DEFAULT_TABLE_PAGE_SIZE }, ...initialState },
-    ...(withPagination && manualPagination
-      ? { manualPagination, onPaginationChange, pageCount }
-      : { getPaginationRowModel: getPaginationRowModel() }),
-    ...(enableSorting && manualSorting
-      ? { manualSorting, onSortingChange }
-      : { getSortedRowModel: getSortedRowModel() }),
-    state: {
-      ...(withPagination && manualPagination ? { pagination: paginationState } : {}),
-      ...(enableSorting && manualSorting ? { sorting: sortingState } : {}),
-      ...(withPagination ? {} : singlePagePaginationState),
-    },
-  });
+	const reactTable = useReactTable({
+		columnResizeMode: 'onChange',
+		columns,
+		data,
+		enableColumnResizing,
+		enableSorting,
+		getCoreRowModel: getCoreRowModel(),
+		initialState: { pagination: { pageSize: DEFAULT_TABLE_PAGE_SIZE }, ...initialState },
+		...(withPagination && manualPagination
+			? { manualPagination, onPaginationChange, pageCount }
+			: { getPaginationRowModel: getPaginationRowModel() }),
+		...(enableSorting && manualSorting
+			? { manualSorting, onSortingChange }
+			: { getSortedRowModel: getSortedRowModel() }),
+		state: {
+			...(withPagination && manualPagination ? { pagination: paginationState } : {}),
+			...(enableSorting && manualSorting ? { sorting: sortingState } : {}),
+			...(withPagination ? {} : singlePagePaginationState),
+		},
+	});
 
-  return (
-    <TableContainer className={className}>
-      <TableContainerInner withFilters={withFilters} withTabs={withTabs}>
-        <TableStyled withSideBorders={withSideBorders}>
-          {withHeaders && (
-            <TableHead>
-              {reactTable.getHeaderGroups().map((headerGroup, headerIndex) => (
-                <TableRow key={headerGroup.id} index={headerIndex} withStripes={withStripes}>
-                  {headerGroup.headers.map((header) => {
-                    const canSort = enableSorting && header.column.getCanSort();
-                    const isCustomHeader = header.column.columnDef.meta?.customHeader;
-                    const headerContents = header.isPlaceholder
-                      ? null
-                      : flexRender(header.column.columnDef.header, header.getContext());
+	return (
+		<TableContainer className={className}>
+			<TableContainerInner withFilters={withFilters} withTabs={withTabs}>
+				<TableStyled withSideBorders={withSideBorders}>
+					{withHeaders && (
+						<TableHead>
+							{reactTable.getHeaderGroups().map((headerGroup, headerIndex) => (
+								<TableRow key={headerGroup.id} index={headerIndex} withStripes={withStripes}>
+									{headerGroup.headers.map((header) => {
+										const canSort = enableSorting && header.column.getCanSort();
+										const isCustomHeader = header.column.columnDef.meta?.customHeader;
+										const headerContents = header.isPlaceholder
+											? null
+											: flexRender(header.column.columnDef.header, header.getContext());
 
-                    const {
-                      activeTab = '',
-                      handleTabs = () => {},
-                      tabs = [],
-                    } = header.column.columnDef.meta?.columnTabs || {};
+										const {
+											activeTab = '',
+											handleTabs = () => {},
+											tabs = [],
+										} = header.column.columnDef.meta?.columnTabs || {};
 
-                    return (
-                      <TableHeader
-                        key={header.id}
-                        colSpan={header.colSpan}
-                        width={header.getSize()}
-                        sorted={header.column.getIsSorted()}
-                        canSort={canSort}
-                      >
-                        {!!tabs.length && (
-                          <TableTabs activeTab={activeTab} handleTabs={handleTabs} tabs={tabs} />
-                        )}
-                        <SortButton
-                          canSort={canSort}
-                          onClick={header.column.getToggleSortingHandler()}
-                        >
-                          {isCustomHeader ? (
-                            headerContents
-                          ) : (
-                            <TableHeaderWrapper>{headerContents}</TableHeaderWrapper>
-                          )}
-                        </SortButton>
-                        {header.column.getCanResize() && (
-                          <Resizer
-                            onMouseDown={header.getResizeHandler()}
-                            onTouchStart={header.getResizeHandler()}
-                            className={`resizer ${
-                              header.column.getIsResizing() ? 'isResizing' : ''
-                            }`}
-                          />
-                        )}
-                      </TableHeader>
-                    );
-                  })}
-                </TableRow>
-              ))}
-            </TableHead>
-          )}
+										return (
+											<TableHeader
+												key={header.id}
+												colSpan={header.colSpan}
+												width={header.getSize()}
+												sorted={header.column.getIsSorted()}
+												canSort={canSort}
+											>
+												{!!tabs.length && (
+													<TableTabs activeTab={activeTab} handleTabs={handleTabs} tabs={tabs} />
+												)}
+												<SortButton
+													canSort={canSort}
+													onClick={header.column.getToggleSortingHandler()}
+												>
+													{isCustomHeader ? (
+														headerContents
+													) : (
+														<TableHeaderWrapper>{headerContents}</TableHeaderWrapper>
+													)}
+												</SortButton>
+												{header.column.getCanResize() && (
+													<Resizer
+														onMouseDown={header.getResizeHandler()}
+														onTouchStart={header.getResizeHandler()}
+														className={`resizer ${
+															header.column.getIsResizing() ? 'isResizing' : ''
+														}`}
+													/>
+												)}
+											</TableHeader>
+										);
+									})}
+								</TableRow>
+							))}
+						</TableHead>
+					)}
 
-          <TableBody>
-            {reactTable.getRowModel().rows.map((row, rowIndex) => (
-              <TableRow
-                index={rowIndex}
-                key={row.id}
-                withRowBorder={withRowBorder}
-                withRowHighlight={withRowHighlight}
-                withStripes={withStripes}
-              >
-                {row.getVisibleCells().map((cell) => {
-                  const isCustomCell = cell.column.columnDef.meta?.customCell;
-                  const cellContents = flexRender(cell.column.columnDef.cell, cell.getContext());
-                  return (
-                    <TableCell key={cell.id} width={cell.column.getSize()}>
-                      {isCustomCell ? (
-                        cellContents
-                      ) : (
-                        <TableCellWrapper>{cellContents}</TableCellWrapper>
-                      )}
-                    </TableCell>
-                  );
-                })}
-              </TableRow>
-            ))}
-          </TableBody>
-        </TableStyled>
-      </TableContainerInner>
-      {withPagination && (
-        <TablePaginationV8
-          canNextPage={reactTable.getCanNextPage()}
-          canPreviousPage={reactTable.getCanPreviousPage()}
-          nextPage={reactTable.nextPage}
-          pageCount={reactTable.getPageCount()}
-          pageIndex={reactTable.getState().pagination.pageIndex}
-          pageSize={reactTable.getState().pagination.pageSize}
-          previousPage={reactTable.previousPage}
-          resetPageIndex={reactTable.resetPageIndex}
-          setPageIndex={reactTable.setPageIndex}
-          setPageSize={reactTable.setPageSize}
-          showPageSizeOptions={showPageSizeOptions}
-        />
-      )}
-      <LoaderComponent $loading={loading} />
-    </TableContainer>
-  );
+					<TableBody>
+						{reactTable.getRowModel().rows.map((row, rowIndex) => (
+							<TableRow
+								index={rowIndex}
+								key={row.id}
+								withRowBorder={withRowBorder}
+								withRowHighlight={withRowHighlight}
+								withStripes={withStripes}
+							>
+								{row.getVisibleCells().map((cell) => {
+									const isCustomCell = cell.column.columnDef.meta?.customCell;
+									const cellContents = flexRender(cell.column.columnDef.cell, cell.getContext());
+									return (
+										<TableCell key={cell.id} width={cell.column.getSize()}>
+											{isCustomCell ? (
+												cellContents
+											) : (
+												<TableCellWrapper>{cellContents}</TableCellWrapper>
+											)}
+										</TableCell>
+									);
+								})}
+							</TableRow>
+						))}
+					</TableBody>
+				</TableStyled>
+			</TableContainerInner>
+			{withPagination && (
+				<TablePaginationV8
+					canNextPage={reactTable.getCanNextPage()}
+					canPreviousPage={reactTable.getCanPreviousPage()}
+					nextPage={reactTable.nextPage}
+					pageCount={reactTable.getPageCount()}
+					pageIndex={reactTable.getState().pagination.pageIndex}
+					pageSize={reactTable.getState().pagination.pageSize}
+					previousPage={reactTable.previousPage}
+					resetPageIndex={reactTable.resetPageIndex}
+					setPageIndex={reactTable.setPageIndex}
+					setPageSize={reactTable.setPageSize}
+					showPageSizeOptions={showPageSizeOptions}
+				/>
+			)}
+			<LoaderComponent $loading={loading} />
+		</TableContainer>
+	);
 };
