+++ src/Tabs/Tabs.stories.tsx
@@ -17,32 +17,49 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-/** @jsxImportSource @emotion/react */
+import { action } from '@storybook/addon-actions';
+import React from 'react';
+import { Tab, Tabs } from '.';
+import { Typography } from '../Typography';
 
-import { css } from '@emotion/react';
-import { boolean, select, text } from '@storybook/addon-knobs';
-import { storiesOf } from '@storybook/react';
-import defaultTheme from '../../theme/defaultTheme';
-import { InteractiveIcon } from './index';
-import { Icons as icons, UikitIconNames } from '../icons';
+function TabContainer(props) {
+	return (
+		<Typography component="div" style={{ padding: 8 * 3 }}>
+			{props.children}
+		</Typography>
+	);
+}
 
-storiesOf(`InteractiveIcon`, module).add('Basic', () => {
-  const knobs = {
-    text: text('value', 'confused?'),
-    fill: select('fill', [null, '#00f', ...Object.keys(defaultTheme.colors)], null),
-    name: select('name', Object.keys(icons) as UikitIconNames[], 'question'),
-    disabled: boolean('disabled', false),
-    position: select('position', ['top', 'bottom', 'left', 'right'], 'bottom'),
-  };
-  return (
-    <div
-      css={css`
-        margin-top: 20px;
-        display: flex;
-        justify-content: center;
-      `}
-    >
-      <InteractiveIcon html={<span>{knobs.text}</span>} {...knobs}></InteractiveIcon>
-    </div>
-  );
+// forwardRef is used to display the source in storybook
+const SimpleTabs = React.forwardRef(() => {
+	const [value, setValue] = React.useState(0);
+
+	function handleChange(event, newValue) {
+		setValue(newValue);
+		action('tab foucs changed')(event, newValue);
+	}
+
+	return (
+		<div>
+			<Tabs value={value} onChange={handleChange}>
+				<Tab value="0" label="Item One" />
+				<Tab value="1" label="Item Two" />
+				<Tab value="2" label="Item Three" />
+				<Tab empty style={{ flexDirection: 'row-reverse' }}>
+					<button>Child of empty tab</button>
+				</Tab>
+			</Tabs>
+			{value === 0 && <TabContainer>Item One</TabContainer>}
+			{value === 1 && <TabContainer>Item Two</TabContainer>}
+			{value === 2 && <TabContainer>Item Three</TabContainer>}
+		</div>
+	);
+});
+
+export default {
+	component: Tabs,
+};
+
+export const Basic = () => ({
+	render: () => <SimpleTabs />,
 });
